show summary line for item cache
git-svn-id: https://svn.exedio.com/cope/trunk/persistence@7329 e7d4fc99-c606-0410-b9bf-843393a9eab7

+++ runtime/src/com/exedio/cope/Cache.java
@@ -271,7 +271,7 @@ final class Cache
 			final long now = System.currentTimeMillis();
 			final int numberOfItemsInCache;
 			long ageSum = 0;
-			long ageMin = Integer.MAX_VALUE;
+			long ageMin = Long.MAX_VALUE;
 			long ageMax = 0;
 
 			final TIntObjectHashMap<PersistentState> stateMap = stateMaps[i];
